build: fall back to bwa-mem2 release version v2.2.1
the dockerized version of this pipeline uses a non-release version of bwa-mem2 on a different
branch, for theoretical speed improvements. however, at least in my hands, the version of the code
that is currently clonable leads to an installation that doesn't work, due to some hardcoded path
wildness. the release version will do for now.