refactor: temporarily swap out s3 remote provider service for direct calls to aws cli
the AWS S3 remote provider functionality in snakemake has some assorted eccentricities. it was more
of a problem the prior time I tried to use it, when it caused horrible lag; but it's still
interfering with acquiring reference files, and so it's out for the time being. I'd like to add it
back eventually.