perf: add performance benchmarks to all burdensome rules
to get sensible tracking files for snakemake benchmarking, some of the patterns are now harmonized
or simplified, to prevent truly terrible subdirectories of results/performance_benchmarks. there are
still a few that will be terrible for the time being; I may go back to clean them up if it becomes
bothersome.