fix: use deprecated cluster config to communicate parallel environment to cluster profile
this is ridiculous, but apparently, as written, the SGE cluster profile has no way to actually pass
in threads data to SGE. this is obviously unacceptable, and so once again I have hacked the profile.
I'll need to post the profile somewhere, now that it's changing so much from default.