fix: improve agreement in format of multiqc report post-alignment
most of this is cleanup to deal with different file paths or names, but one interesting thing: it
seems like the giant block of configuration settings for alignstats is just being completely
ignored, as it doesn't respond to changing the settings at all. this is likely flagged for removal,
and thank goodness, because it's an ugly mess.