feat: implement basic ensemble sv calling
svdb merge can annotate the vcf info file for each record with a pipe-delimited set of the source
files in which the variant (or its overlapping friends) was detected. this information is now used
as a simple way to select out the number of source files for a final ensemble variant. the output
filenames from duphold are modified so that they unambiguously list source information in the
filename itself, as opposed to in the path, which svdb strips. minimum ensemble source count is now
userspace-configurable