feat: add logic to construct basic sample ID linker from lab logbook
this logic is pretty basic, and requires the user to specify a logbook download from the google
drive version of the thing that's maintained by lab. there are many format deviations, but I've
inspected the results for all samples run to date on actual novaseq runs, and it works unambiguously
for all libraries that don't correspond to GIAB references. not sure what to do about GIAB. the
convention is evidently to run a low depth pass of NA24385/HG002 as the last (lexicographically)
sample in the flow cell, but in one flow cell they ran extras of NA12878 and NA24385, and there
doesn't seem to be an explicit annotation of this fact anywhere in the logbook.