feat: change alignstats combined output format from tsv to multiqc-specific yaml
I saw some discrepancies in the multiqc custom tabular output for alignstats when the data were
provided in tsv format. inspection of the multiqc source code shows that evidently certain features
are only supported for custom content when the custom content is provided in json or yaml format.
there doesn't appear to be any particular requirement for this to be so, but it's how it is, and
it's easy enough to swap out