build: adjust R scripts to function with publicly maintained docker image
the rocker::tidyverse R image is great, and I've had good success with it in the past. however, it
likely lacks certain packages these scripts use. as such, I've wrapped some install.packages
statements in the scripts to do a one-time installation via CRAN of missing packages. it's a bit of
a pain, but better than maintaining a boutique R container