feat: expose bwa -K parameter to userspace config
fixing the K parameter to bwa/bwa-mem2 is important for reproducible behavior, as its default is to
scale with the number of threads available, which may shift around. however, there are potentially
performance costs to particular settings of K. the bwa default, setting K to 1e7*{threads}, costs
somewhat more RAM than necessary but potentially improves runtime by 20% or more (relative to the
setting that had been used previously). the RAM cost is somewhat marginal, relative to the truly
remarkable amount of RAM bwa requires regardless. this information was compiled specifically from
bwa, not bwa-mem2, so ymmv