build: assume aws cli is present in global environment
this may not even remotely be a safe assumption. it appears that the environment doesn't inherit
cached credentials in containers, which makes sense. however, since the pipeline assumes cached
environment credentials to avoid security exposure, this is kind of a problem. for the time being,
I'm moving to this assumption, but I think this will have to be changed