feat: allow user config control of whether pipeline chooses containers or conda envs
this is a modification of the standard snakemake behavior for conda environments and containers. if
conda envs and containers were available for all relevant rules, this behavior could be controlled
on the command line with --use-conda or --use-singularity. however, deepvariant does not have a
functional conda environment, and as such always needs to choose its docker container; and so
--use-singularity must always be active. this causes issues when snakemake then uses its default
behavior of launching the container environment and then additionally building a conda environment
within it. now, with some admittedly verbose syntax, a user configuration variable 'use-containers'
causes the workflow to select the rule specific container to the exclusion of the conda environment,
or vice versa.