fix: explicitly flag checkpoint output as input to rule
this shouldn't technically be required, but we've evidently hit the invisible ceiling on the level
of permitted complexity of checkpoints in a workflow. without this input, the checkpoint is not
completing before the downstream rule, leading to undefined behavior. with this input, the
checkpoint at least has to complete, but it's not necessarily guaranteed to work? it's also possible
the temp flag on the checkpoint output was the problem, which is bizarre in its own way